Added use of blob storage. Major Commit
Now authentication is performed using User datatype gained via credentials.json

Will need to work on simplifying it, may change the json to create the type to begin with.

Will need to refactor a bit but it is working well on requesting access versus able to login.

Will need to cut bloat.

Would be nice to see about passwordless sign on for request access versus login,